change int64_t to uint32_t used by flatbuffers

+++ src/common/logging.cc
@@ -90,9 +90,9 @@ void RayLogger_log(RayLogger *logger,
 
 void RayLogger_log_event(DBHandle *db,
                          uint8_t *key,
-                         int64_t key_length,
+                         uint32_t key_length,
                          uint8_t *value,
-                         int64_t value_length,
+                         uint32_t value_length,
                          double timestamp) {
   std::string timestamp_string = std::to_string(timestamp);
   int status = redisAsyncCommand(db->context, NULL, NULL, "ZADD %b %s %b", key,
